feat: simulator and noir for milestone 1.4
