+++ yarn-project/acir-simulator/src/acvm/index.ts
@@ -1,0 +1,3 @@
+export * from './fields.js';
+export * from './acvm.js';
+export * from './witness_io.js';
