+++ yarn-project/acir-simulator/src/arguments_encoder/index.ts
