+++ yarn-project/acir-simulator/src/execution.ts
@@ -1,13 +1,34 @@
-import { ACVMField, acvm, toACVMField, fromACVMField, ZERO_ACVM_FIELD } from './acvm.js';
-import { AztecAddress, CallContext, EthAddress, Fr, OldTreeRoots, TxRequest } from '@aztec/circuits.js';
-import { DBOracle, PrivateCallStackItem } from './db_oracle.js';
-import { writeInputs, extractPublicInputs, frToAztecAddress } from './witness_io.js';
+import { ACVMField, acvm, toACVMField, fromACVMField, ZERO_ACVM_FIELD } from './acvm/index.js';
+import {
+  AztecAddress,
+  CallContext,
+  EthAddress,
+  Fr,
+  OldTreeRoots,
+  TxRequest,
+  PrivateCallStackItem,
+  FunctionData,
+} from '@aztec/circuits.js';
+import { DBOracle } from './db_oracle.js';
+import { writeInputs, extractPublicInputs, frToAztecAddress } from './acvm/witness_io.js';
 import { FunctionAbi } from '@aztec/noir-contracts';
-import { encodeArguments } from './arguments.js';
+import { encodeArguments } from './arguments_encoder/index.js';
+
+interface NewNoteData {
+  preimage: Fr[];
+  storageSlot: Fr;
+  owner: { x: Fr; y: Fr };
+}
+
+interface NewNullifierData {
+  preimage: Fr[];
+  storageSlot: Fr;
+  nullifier: Fr;
+}
 
 export interface ExecutionPreimages {
-  newNotes: Array<{ preimage: Fr[]; storageSlot: Fr }>;
-  nullifiedNotes: Fr[];
+  newNotes: NewNoteData[];
+  nullifiedNotes: NewNullifierData[];
 }
 
 export interface ExecutionResult {
@@ -48,7 +69,7 @@ export class Execution {
     return this.runExternalFunction(
       this.entryPointABI,
       this.contractAddress,
-      this.request.functionData.functionSelector,
+      this.request.functionData,
       encodedArgs,
       callContext,
     );
@@ -58,14 +79,14 @@ export class Execution {
   private async runExternalFunction(
     abi: FunctionAbi,
     contractAddress: AztecAddress,
-    functionSelector: Buffer,
+    functionData: FunctionData,
     args: Fr[],
     callContext: CallContext,
   ): Promise<ExecutionResult> {
     const acir = Buffer.from(abi.bytecode, 'hex');
     const initialWitness = writeInputs(args, callContext, this.request.txContext, this.oldRoots);
-    const newNotePreimages: Array<{ preimage: Fr[]; storageSlot: Fr }> = [];
-    const newNullifiers: Fr[] = [];
+    const newNotePreimages: NewNoteData[] = [];
+    const newNullifiers: NewNullifierData[] = [];
     const nestedExecutionContexts: ExecutionResult[] = [];
 
     const { partialWitness } = await acvm(acir, initialWitness, {
@@ -76,24 +97,30 @@ export class Execution {
         return await this.getNotes(contractAddress, storageSlot, 2);
       },
       getRandomField: () => Promise.resolve([toACVMField(Fr.random())]),
-      notifyCreatedNote: (params: ACVMField[]) => {
-        const [storageSlot, ...acvmPreimage] = params;
-        const preimage = acvmPreimage.map(f => fromACVMField(f));
+      notifyCreatedNote: ([storageSlot, ownerX, ownerY, ...acvmPreimage]: ACVMField[]) => {
         newNotePreimages.push({
-          preimage,
+          preimage: acvmPreimage.map(f => fromACVMField(f)),
           storageSlot: fromACVMField(storageSlot),
+          owner: {
+            x: fromACVMField(ownerX),
+            y: fromACVMField(ownerY),
+          },
         });
         return Promise.resolve([ZERO_ACVM_FIELD]);
       },
-      notifyNullifiedNote: ([nullifier]: ACVMField[]) => {
-        newNullifiers.push(fromACVMField(nullifier));
+      notifyNullifiedNote: ([slot, nullifier, ...acvmPreimage]: ACVMField[]) => {
+        newNullifiers.push({
+          preimage: acvmPreimage.map(f => fromACVMField(f)),
+          storageSlot: fromACVMField(slot),
+          nullifier: fromACVMField(nullifier),
+        });
         return Promise.resolve([ZERO_ACVM_FIELD]);
       },
     });
 
     const publicInputs = extractPublicInputs(partialWitness, acir);
 
-    const callStackItem = new PrivateCallStackItem(contractAddress, functionSelector, publicInputs);
+    const callStackItem = new PrivateCallStackItem(contractAddress, functionData, publicInputs);
 
     return {
       acir,
