+++ yarn-project/acir-simulator/src/index.test.ts
@@ -1,5 +1,4 @@
 import {
-  ARGS_LENGTH,
   AztecAddress,
   ContractDeploymentData,
   EthAddress,
