+++ yarn-project/acir-simulator/src/index.ts
@@ -1,4 +1,4 @@
 export * from './simulator.js';
 export * from './execution.js';
 export * from './db_oracle.js';
-export * from './acvm.js';
+export * from './acvm/acvm.js';
